@@ -1,191 +1,191 @@
 <div class="nk-footer nk-auth-footer-full">
-                        <div class="container wide-lg">
-                            <div class="row g-3">
-                                <div class="col-lg-6 order-lg-last">
-                                    <ul class="nav nav-sm justify-content-center justify-content-lg-end">
-                                        <li class="nav-item">
-                                            <a class="nav-link" href="#">Terms & Condition</a>
-                                        </li>
-                                        <li class="nav-item">
-                                            <a class="nav-link" href="#">Privacy Policy</a>
-                                        </li>
-                                        <li class="nav-item">
-                                            <a class="nav-link" href="#">Help</a>
-                                        </li>
-                                        <li class="nav-item dropup">
-                                            <a class="dropdown-toggle dropdown-indicator has-indicator nav-link" data-toggle="dropdown" data-offset="0,10"><span>English</span></a>
-                                            <div class="dropdown-menu dropdown-menu-sm dropdown-menu-right">
-                                                <ul class="language-list">
-                                                    <li>
-                                                        <a href="#" class="language-item">
-                                                            <img src="<?php echo base_url()?>/files/images/flags/english.png" alt="" class="language-flag">
-                                                            <span class="language-name">English</span>
-                                                        </a>
-                                                    </li>
-                                                    <li>
-                                                        <a href="#" class="language-item">
-                                                            <img src="<?php echo base_url()?>/files/images/flags/spanish.png" alt="" class="language-flag">
-                                                            <span class="language-name">Español</span>
-                                                        </a>
-                                                    </li>
-                                                    <li>
-                                                        <a href="#" class="language-item">
-                                                            <img src="<?php echo base_url()?>/files/images/flags/french.png" alt="" class="language-flag">
-                                                            <span class="language-name">Français</span>
-                                                        </a>
-                                                    </li>
-                                                    <li>
-                                                        <a href="#" class="language-item">
-                                                            <img src="<?php echo base_url()?>/files/images/flags/turkey.png" alt="" class="language-flag">
-                                                            <span class="language-name">Türkçe</span>
-                                                        </a>
-                                                    </li>
-                                                </ul>
-                                            </div>
-                                        </li>
-                                    </ul>
-                                </div>
-                                <div class="col-lg-6">
-                                    <div class="nk-block-content text-center text-lg-left">
-                                        <p class="text-soft">&copy; 2022 Camera20 Productions. All Rights Reserved.</p>
-                                    </div>
-                                </div>
-                            </div>
+    <div class="container wide-lg">
+        <div class="row g-3">
+            <div class="col-lg-6 order-lg-last">
+                <ul class="nav nav-sm justify-content-center justify-content-lg-end">
+                    <li class="nav-item">
+                        <a class="nav-link" href="#">Terms & Condition</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="#">Privacy Policy</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="#">Help</a>
+                    </li>
+                    <li class="nav-item dropup">
+                        <a class="dropdown-toggle dropdown-indicator has-indicator nav-link" data-toggle="dropdown" data-offset="0,10"><span>English</span></a>
+                        <div class="dropdown-menu dropdown-menu-sm dropdown-menu-right">
+                            <ul class="language-list">
+                                <li>
+                                    <a href="#" class="language-item">
+                                        <img src="<?php echo base_url() ?>/files/images/flags/english.png" alt="" class="language-flag">
+                                        <span class="language-name">English</span>
+                                    </a>
+                                </li>
+                                <li>
+                                    <a href="#" class="language-item">
+                                        <img src="<?php echo base_url() ?>/files/images/flags/spanish.png" alt="" class="language-flag">
+                                        <span class="language-name">Español</span>
+                                    </a>
+                                </li>
+                                <li>
+                                    <a href="#" class="language-item">
+                                        <img src="<?php echo base_url() ?>/files/images/flags/french.png" alt="" class="language-flag">
+                                        <span class="language-name">Français</span>
+                                    </a>
+                                </li>
+                                <li>
+                                    <a href="#" class="language-item">
+                                        <img src="<?php echo base_url() ?>/files/images/flags/turkey.png" alt="" class="language-flag">
+                                        <span class="language-name">Türkçe</span>
+                                    </a>
+                                </li>
+                            </ul>
                         </div>
-                    </div>
-                </div>
-                <!-- wrap @e -->
+                    </li>
+                </ul>
             </div>
-            <!-- content @e -->
-        </div>
-        <!-- main @e -->
-    </div>
-    <!-- app-root @e -->
-    <!-- JavaScript -->
-    <script src="<?php echo base_url()?>/files/assets/js/bundle.js?ver=2.9.1"></script>
-    <script src="<?php echo base_url()?>/files/assets/js/scripts.js?ver=2.9.1"></script>
-    <!-- select region modal -->
-    <div class="modal fade" tabindex="-1" role="dialog" id="region">
-        <div class="modal-dialog modal-lg" role="document">
-            <div class="modal-content">
-                <a href="#" class="close" data-dismiss="modal"><em class="icon ni ni-cross-sm"></em></a>
-                <div class="modal-body modal-body-md">
-                    <h5 class="title mb-4">Select Your Country</h5>
-                    <div class="nk-country-region">
-                        <ul class="country-list text-center gy-2">
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/arg.png" alt="" class="country-flag">
-                                    <span class="country-name">Argentina</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/aus.png" alt="" class="country-flag">
-                                    <span class="country-name">Australia</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/bangladesh.png" alt="" class="country-flag">
-                                    <span class="country-name">Bangladesh</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/canada.png" alt="" class="country-flag">
-                                    <span class="country-name">Canada <small>(English)</small></span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/china.png" alt="" class="country-flag">
-                                    <span class="country-name">Centrafricaine</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/china.png" alt="" class="country-flag">
-                                    <span class="country-name">China</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/french.png" alt="" class="country-flag">
-                                    <span class="country-name">France</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/germany.png" alt="" class="country-flag">
-                                    <span class="country-name">Germany</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/iran.png" alt="" class="country-flag">
-                                    <span class="country-name">Iran</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/italy.png" alt="" class="country-flag">
-                                    <span class="country-name">Italy</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/mexico.png" alt="" class="country-flag">
-                                    <span class="country-name">México</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/philipine.png" alt="" class="country-flag">
-                                    <span class="country-name">Philippines</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/portugal.png" alt="" class="country-flag">
-                                    <span class="country-name">Portugal</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/s-africa.png" alt="" class="country-flag">
-                                    <span class="country-name">South Africa</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/spanish.png" alt="" class="country-flag">
-                                    <span class="country-name">Spain</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/switzerland.png" alt="" class="country-flag">
-                                    <span class="country-name">Switzerland</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/uk.png" alt="" class="country-flag">
-                                    <span class="country-name">United Kingdom</span>
-                                </a>
-                            </li>
-                            <li>
-                                <a href="#" class="country-item">
-                                    <img src="<?php echo base_url()?>/files/images/flags/english.png" alt="" class="country-flag">
-                                    <span class="country-name">United State</span>
-                                </a>
-                            </li>
-                        </ul>
-                    </div>
+            <div class="col-lg-6">
+                <div class="nk-block-content text-center text-lg-left">
+                    <p class="text-soft">&copy; 2022 - <?php echo date("Y"); ?> Camera20 Productions. All Rights Reserved.</p>
                 </div>
-            </div><!-- .modal-content -->
-        </div><!-- .modla-dialog -->
-    </div><!-- .modal -->
+            </div>
+        </div>
+    </div>
+</div>
+</div>
+<!-- wrap @e -->
+</div>
+<!-- content @e -->
+</div>
+<!-- main @e -->
+</div>
+<!-- app-root @e -->
+<!-- JavaScript -->
+<script src="<?php echo base_url() ?>/files/assets/js/bundle.js?ver=2.9.1"></script>
+<script src="<?php echo base_url() ?>/files/assets/js/scripts.js?ver=2.9.1"></script>
+<!-- select region modal -->
+<div class="modal fade" tabindex="-1" role="dialog" id="region">
+    <div class="modal-dialog modal-lg" role="document">
+        <div class="modal-content">
+            <a href="#" class="close" data-dismiss="modal"><em class="icon ni ni-cross-sm"></em></a>
+            <div class="modal-body modal-body-md">
+                <h5 class="title mb-4">Select Your Country</h5>
+                <div class="nk-country-region">
+                    <ul class="country-list text-center gy-2">
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/arg.png" alt="" class="country-flag">
+                                <span class="country-name">Argentina</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/aus.png" alt="" class="country-flag">
+                                <span class="country-name">Australia</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/bangladesh.png" alt="" class="country-flag">
+                                <span class="country-name">Bangladesh</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/canada.png" alt="" class="country-flag">
+                                <span class="country-name">Canada <small>(English)</small></span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/china.png" alt="" class="country-flag">
+                                <span class="country-name">Centrafricaine</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/china.png" alt="" class="country-flag">
+                                <span class="country-name">China</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/french.png" alt="" class="country-flag">
+                                <span class="country-name">France</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/germany.png" alt="" class="country-flag">
+                                <span class="country-name">Germany</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/iran.png" alt="" class="country-flag">
+                                <span class="country-name">Iran</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/italy.png" alt="" class="country-flag">
+                                <span class="country-name">Italy</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/mexico.png" alt="" class="country-flag">
+                                <span class="country-name">México</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/philipine.png" alt="" class="country-flag">
+                                <span class="country-name">Philippines</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/portugal.png" alt="" class="country-flag">
+                                <span class="country-name">Portugal</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/s-africa.png" alt="" class="country-flag">
+                                <span class="country-name">South Africa</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/spanish.png" alt="" class="country-flag">
+                                <span class="country-name">Spain</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/switzerland.png" alt="" class="country-flag">
+                                <span class="country-name">Switzerland</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/uk.png" alt="" class="country-flag">
+                                <span class="country-name">United Kingdom</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#" class="country-item">
+                                <img src="<?php echo base_url() ?>/files/images/flags/english.png" alt="" class="country-flag">
+                                <span class="country-name">United State</span>
+                            </a>
+                        </li>
+                    </ul>
+                </div>
+            </div>
+        </div><!-- .modal-content -->
+    </div><!-- .modla-dialog -->
+</div><!-- .modal -->
 
 </html>
