@@ -1,460 +1,315 @@
 <!-- content @s -->
-<div class="nk-content ">
-                <div class="container-fluid">
-                    <div class="nk-content-inner">
-                        <div class="nk-content-body">
-                            <div class="nk-block-head nk-block-head-sm">
-                                <div class="nk-block-between">
-                                    <div class="nk-block-head-content">
-                                        <h3 class="nk-block-title page-title">Users Lists</h3>
-                                        <div class="nk-block-des text-soft">
-                                            <p>You have total  users.</p>
-                                        </div>
-                                    </div><!-- .nk-block-head-content -->
-                                    <div class="nk-block-head-content">
-                                        <div class="toggle-wrap nk-block-tools-toggle">
-                                            <a href="/html/user-list-regular.html" class="btn btn-icon btn-trigger toggle-expand mr-n1" data-target="pageMenu"><em class="icon ni ni-menu-alt-r"></em></a>
-                                            <div class="toggle-expand-content" data-content="pageMenu">
-                                                <ul class="nk-block-tools g-3">
-                                                    <li><a href="/html/user-list-regular.html" class="btn btn-white btn-outline-light"><em class="icon ni ni-download-cloud"></em><span>Export</span></a></li>
-                                                    <li class="nk-block-tools-opt">
-                                                    <li class="nk-block-tools-opt">
-                                                        <a href="/html/user-list-regular.html" data-target="addProduct" class="toggle btn btn-icon btn-primary d-md-none"><em class="icon ni ni-plus"></em></a>
-                                                        <a href="/html/user-list-regular.html" data-target="addProduct" class="toggle btn btn-primary d-none d-md-inline-flex"><em class="icon ni ni-plus"></em><span>Add User</span></a>
-                                                    </li>
-                                                    </li>
-                                                </ul>
-                                            </div>
-                                        </div><!-- .toggle-wrap -->
-                                    </div><!-- .nk-block-head-content -->
-                                </div><!-- .nk-block-between -->
-                            </div><!-- .nk-block-head -->
-                            <div class="nk-block">
-                                <div class="card card-bordered card-stretch">
-                                    <div class="card-inner-group">
-                                        <div class="card-inner position-relative card-tools-toggle">
-                                            <div class="card-title-group">
-                                                <div class="card-tools">
-                                                    <div class="form-inline flex-nowrap gx-3">
-                                                        <div class="form-wrap w-150px">
-                                                            <select class="form-select" data-search="off" data-placeholder="Bulk Action">
-                                                                <option value="">Bulk Action</option>
-                                                                <option value="email">Send Email</option>
-                                                                <option value="group">Change Group</option>
-                                                                <option value="suspend">Suspend User</option>
-                                                                <option value="delete">Delete User</option>
-                                                            </select>
-                                                        </div>
-                                                        <div class="btn-wrap">
-                                                            <span class="d-none d-md-block"><button class="btn btn-dim btn-outline-light disabled">Apply</button></span>
-                                                            <span class="d-md-none"><button class="btn btn-dim btn-outline-light btn-icon disabled"><em class="icon ni ni-arrow-right"></em></button></span>
-                                                        </div>
-                                                    </div><!-- .form-inline -->
-                                                </div><!-- .card-tools -->
-                                                <div class="card-tools mr-n1">
-                                                    <ul class="btn-toolbar gx-1">
-                                                        <li>
-                                                            <a href="#" class="btn btn-icon search-toggle toggle-search" data-target="search"><em class="icon ni ni-search"></em></a>
-                                                        </li><!-- li -->
-                                                        <li class="btn-toolbar-sep"></li><!-- li -->
-                                                        <li>
-                                                            <div class="toggle-wrap">
-                                                                <a href="/html/user-list-regular.html" class="btn btn-icon btn-trigger toggle" data-target="cardTools"><em class="icon ni ni-menu-right"></em></a>
-                                                                <div class="toggle-content" data-content="cardTools">
-                                                                    <ul class="btn-toolbar gx-1">
-                                                                        <li class="toggle-close">
-                                                                            <a href="/html/user-list-regular.html" class="btn btn-icon btn-trigger toggle" data-target="cardTools"><em class="icon ni ni-arrow-left"></em></a>
-                                                                        </li><!-- li -->
-                                                                        <li>
-                                                                            <div class="dropdown">
-                                                                                <a href="#" class="btn btn-trigger btn-icon dropdown-toggle" data-toggle="dropdown">
-                                                                                    <div class="dot dot-primary"></div>
-                                                                                    <em class="icon ni ni-filter-alt"></em>
-                                                                                </a>
-                                                                                <div class="filter-wg dropdown-menu dropdown-menu-xl dropdown-menu-right">
-                                                                                    <div class="dropdown-head">
-                                                                                        <span class="sub-title dropdown-title">Filter Users</span>
-                                                                                        <div class="dropdown">
-                                                                                            <a href="#" class="btn btn-sm btn-icon">
-                                                                                                <em class="icon ni ni-more-h"></em>
-                                                                                            </a>
-                                                                                        </div>
-                                                                                    </div>
-                                                                                    <div class="dropdown-body dropdown-body-rg">
-                                                                                        <div class="row gx-6 gy-3">
-                                                                                            <div class="col-6">
-                                                                                                <div class="custom-control custom-control-sm custom-checkbox">
-                                                                                                    <input type="checkbox" class="custom-control-input" id="hasBalance">
-                                                                                                    <label class="custom-control-label" for="hasBalance"> Have Balance</label>
-                                                                                                </div>
-                                                                                            </div>
-                                                                                            <div class="col-6">
-                                                                                                <div class="custom-control custom-control-sm custom-checkbox">
-                                                                                                    <input type="checkbox" class="custom-control-input" id="hasKYC">
-                                                                                                    <label class="custom-control-label" for="hasKYC"> KYC Verified</label>
-                                                                                                </div>
-                                                                                            </div>
-                                                                                            <div class="col-6">
-                                                                                                <div class="form-group">
-                                                                                                    <label class="overline-title overline-title-alt">Role</label>
-                                                                                                    <select class="form-select">
-                                                                                                        <option value="any">Any Role</option>
-                                                                                                        <option value="investor">Investor</option>
-                                                                                                        <option value="seller">Seller</option>
-                                                                                                        <option value="buyer">Buyer</option>
-                                                                                                    </select>
-                                                                                                </div>
-                                                                                            </div>
-                                                                                            <div class="col-6">
-                                                                                                <div class="form-group">
-                                                                                                    <label class="overline-title overline-title-alt">Status</label>
-                                                                                                    <select class="form-select">
-                                                                                                        <option value="any">Any Status</option>
-                                                                                                        <option value="active">Active</option>
-                                                                                                        <option value="pending">Pending</option>
-                                                                                                        <option value="suspend">Suspend</option>
-                                                                                                        <option value="deleted">Deleted</option>
-                                                                                                    </select>
-                                                                                                </div>
-                                                                                            </div>
-                                                                                            <div class="col-12">
-                                                                                                <div class="form-group">
-                                                                                                    <button type="button" class="btn btn-secondary">Filter</button>
-                                                                                                </div>
-                                                                                            </div>
-                                                                                        </div>
-                                                                                    </div>
-                                                                                    <div class="dropdown-foot between">
-                                                                                        <a class="clickable" href="/html/user-list-regular.html">Reset Filter</a>
-                                                                                        <a href="/html/user-list-regular.html">Save Filter</a>
-                                                                                    </div>
-                                                                                </div><!-- .filter-wg -->
-                                                                            </div><!-- .dropdown -->
-                                                                        </li><!-- li -->
-                                                                        <li>
-                                                                            <div class="dropdown">
-                                                                                <a href="/html/user-list-regular.html" class="btn btn-trigger btn-icon dropdown-toggle" data-toggle="dropdown">
-                                                                                    <em class="icon ni ni-setting"></em>
-                                                                                </a>
-                                                                                <div class="dropdown-menu dropdown-menu-xs dropdown-menu-right">
-                                                                                    <ul class="link-check">
-                                                                                        <li><span>Show</span></li>
-                                                                                        <li class="active"><a href="/html/user-list-regular.html">10</a></li>
-                                                                                        <li><a href="/html/user-list-regular.html">20</a></li>
-                                                                                        <li><a href="/html/user-list-regular.html">50</a></li>
-                                                                                    </ul>
-                                                                                    <ul class="link-check">
-                                                                                        <li><span>Order</span></li>
-                                                                                        <li class="active"><a href="/html/user-list-regular.html">DESC</a></li>
-                                                                                        <li><a href="/html/user-list-regular.html">ASC</a></li>
-                                                                                    </ul>
-                                                                                </div>
-                                                                            </div><!-- .dropdown -->
-                                                                        </li><!-- li -->
-                                                                    </ul><!-- .btn-toolbar -->
-                                                                </div><!-- .toggle-content -->
-                                                            </div><!-- .toggle-wrap -->
-                                                        </li><!-- li -->
-                                                    </ul><!-- .btn-toolbar -->
-                                                </div><!-- .card-tools -->
-                                            </div><!-- .card-title-group -->
-                                            <div class="card-search search-wrap" data-search="search">
-                                                <div class="card-body">
-                                                    <div class="search-content">
-                                                        <a href="/html/user-list-regular.html" class="search-back btn btn-icon toggle-search" data-target="search"><em class="icon ni ni-arrow-left"></em></a>
-                                                        <input type="text" class="form-control border-transparent form-focus-none" placeholder="Search by user or email">
-                                                        <button class="search-submit btn btn-icon"><em class="icon ni ni-search"></em></button>
-                                                    </div>
-                                                </div>
-                                            </div><!-- .card-search -->
-                                        </div><!-- .card-inner -->
-                                        <div class="card-inner p-0">
-                                            <div class="nk-tb-list nk-tb-ulist">
-                                                <div class="nk-tb-item nk-tb-head">
-                                                    <div class="nk-tb-col nk-tb-col-check">
-                                                        <div class="custom-control custom-control-sm custom-checkbox notext">
-                                                            <input type="checkbox" class="custom-control-input" id="uid">
-                                                            <label class="custom-control-label" for="uid"></label>
-                                                        </div>
-                                                    </div>
-                                                    <div class="nk-tb-col"><span class="sub-text">User</span></div>
-                                                    <div class="nk-tb-col tb-col-mb"><span class="sub-text">ID Number</span></div>
-                                                    <div class="nk-tb-col tb-col-md"><span class="sub-text">Phone</span></div>
-                                                    <div class="nk-tb-col tb-col-lg"><span class="sub-text">Verified</span></div>
-                                                    <div class="nk-tb-col tb-col-lg"><span class="sub-text">Added On</span></div>
-                                                    <div class="nk-tb-col tb-col-md"><span class="sub-text">Account Type</span></div>
-                                                    <div class="nk-tb-col nk-tb-col-tools text-right">
-                                                        <div class="dropdown">
-                                                            <a href="/html/user-list-regular.html" class="btn btn-xs btn-outline-light btn-icon dropdown-toggle" data-toggle="dropdown" data-offset="0,5"><em class="icon ni ni-plus"></em></a>
-                                                            <div class="dropdown-menu dropdown-menu-xs dropdown-menu-right">
-                                                                <ul class="link-tidy sm no-bdr">
-                                                                    <li>
-                                                                        <div class="custom-control custom-control-sm custom-checkbox">
-                                                                            <input type="checkbox" class="custom-control-input" checked="" id="bl">
-                                                                            <label class="custom-control-label" for="bl">ID Number</label>
-                                                                        </div>
-                                                                    </li>
-                                                                    <li>
-                                                                        <div class="custom-control custom-control-sm custom-checkbox">
-                                                                            <input type="checkbox" class="custom-control-input" checked="" id="ph">
-                                                                            <label class="custom-control-label" for="ph">Phone</label>
-                                                                        </div>
-                                                                    </li>
-                                                                    <li>
-                                                                        <div class="custom-control custom-control-sm custom-checkbox">
-                                                                            <input type="checkbox" class="custom-control-input" id="vri">
-                                                                            <label class="custom-control-label" for="vri">Verified</label>
-                                                                        </div>
-                                                                    </li>
-                                                                    <li>
-                                                                        <div class="custom-control custom-control-sm custom-checkbox">
-                                                                            <input type="checkbox" class="custom-control-input" id="st">
-                                                                            <label class="custom-control-label" for="st">Status</label>
-                                                                        </div>
-                                                                    </li>
-                                                                </ul>
-                                                            </div>
-                                                        </div>
-                                                    </div>
-                                                </div><!-- .nk-tb-item -->
-                                                
+<?= $this->extend('1layouts/main') ?>
 
-                                                <?php foreach($user as $row):?> 
-                                                <div class="nk-tb-item">
-                                                    <div class="nk-tb-col nk-tb-col-check">
-                                                        <div class="custom-control custom-control-sm custom-checkbox notext">
-                                                            <input type="checkbox" class="custom-control-input" id="uid10">
-                                                            <label class="custom-control-label" for="uid10"></label>
-                                                        </div>
-                                                    </div>
-                                                    <div class="nk-tb-col">
-                                                        <div class="user-card">
-                                                            <div class="user-avatar bg-purple">
-                                                                <span><?php echo substr($row['user_fname'],0,1) ?><?php echo substr($row['user_oname'],0,1) ?></span>
+<?= $this->section('cp') ?>
+
+
+
+<div class="nk-content-body">
+    <div class="nk-block-head nk-block-head-sm">
+        <div class="nk-block-between">
+            <div class="nk-block-head-content">
+                <h3 class="nk-block-title page-title">Users Lists</h3>
+                <div class="nk-block-des text-soft">
+                    <p>You have total users.</p>
+                </div>
+            </div><!-- .nk-block-head-content -->
+            <div class="nk-block-head-content">
+                <div class="toggle-wrap nk-block-tools-toggle">
+                    <a href="/html/user-list-regular.html" class="btn btn-icon btn-trigger toggle-expand mr-n1" data-target="pageMenu"><em class="icon ni ni-menu-alt-r"></em></a>
+                    <div class="toggle-expand-content" data-content="pageMenu">
+                        <ul class="nk-block-tools g-3">
+                            <li><a href="/html/user-list-regular.html" class="btn btn-white btn-outline-light"><em class="icon ni ni-download-cloud"></em><span>Export</span></a></li>
+                            <li class="nk-block-tools-opt">
+                            <li class="nk-block-tools-opt">
+                                <a href="/html/user-list-regular.html" data-target="addProduct" class="toggle btn btn-icon btn-primary d-md-none"><em class="icon ni ni-plus"></em></a>
+                                <a href="/html/user-list-regular.html" data-target="addProduct" class="toggle btn btn-primary d-none d-md-inline-flex"><em class="icon ni ni-plus"></em><span>Add User</span></a>
+                            </li>
+                            </li>
+                        </ul>
+                    </div>
+                </div><!-- .toggle-wrap -->
+            </div><!-- .nk-block-head-content -->
+        </div><!-- .nk-block-between -->
+    </div><!-- .nk-block-head -->
+    <div class="nk-block">
+        <div class="card card-bordered card-stretch">
+            <div class="card-inner-group">
+                <div class="card-inner position-relative card-tools-toggle">
+                    <div class="card-title-group">
+                        <div class="card-tools">
+                            <div class="form-inline flex-nowrap gx-3">
+                                <div class="form-wrap w-150px">
+                                    <select class="form-select" data-search="off" data-placeholder="Bulk Action">
+                                        <option value="">Bulk Action</option>
+                                        <option value="email">Send Email</option>
+                                        <option value="group">Change Group</option>
+                                        <option value="suspend">Suspend User</option>
+                                        <option value="delete">Delete User</option>
+                                    </select>
+                                </div>
+                                <div class="btn-wrap">
+                                    <span class="d-none d-md-block"><button class="btn btn-dim btn-outline-light disabled">Apply</button></span>
+                                    <span class="d-md-none"><button class="btn btn-dim btn-outline-light btn-icon disabled"><em class="icon ni ni-arrow-right"></em></button></span>
+                                </div>
+                            </div><!-- .form-inline -->
+                        </div><!-- .card-tools -->
+                        <div class="card-tools mr-n1">
+                            <ul class="btn-toolbar gx-1">
+                                <li>
+                                    <a href="#" class="btn btn-icon search-toggle toggle-search" data-target="search"><em class="icon ni ni-search"></em></a>
+                                </li><!-- li -->
+                                <li class="btn-toolbar-sep"></li><!-- li -->
+                                <li>
+                                    <div class="toggle-wrap">
+                                        <a href="/html/user-list-regular.html" class="btn btn-icon btn-trigger toggle" data-target="cardTools"><em class="icon ni ni-menu-right"></em></a>
+                                        <div class="toggle-content" data-content="cardTools">
+                                            <ul class="btn-toolbar gx-1">
+                                                <li class="toggle-close">
+                                                    <a href="/html/user-list-regular.html" class="btn btn-icon btn-trigger toggle" data-target="cardTools"><em class="icon ni ni-arrow-left"></em></a>
+                                                </li><!-- li -->
+                                                <li>
+                                                    <div class="dropdown">
+                                                        <a href="#" class="btn btn-trigger btn-icon dropdown-toggle" data-toggle="dropdown">
+                                                            <div class="dot dot-primary"></div>
+                                                            <em class="icon ni ni-filter-alt"></em>
+                                                        </a>
+                                                        <div class="filter-wg dropdown-menu dropdown-menu-xl dropdown-menu-right">
+                                                            <div class="dropdown-head">
+                                                                <span class="sub-title dropdown-title">Filter Users</span>
+                                                                <div class="dropdown">
+                                                                    <a href="#" class="btn btn-sm btn-icon">
+                                                                        <em class="icon ni ni-more-h"></em>
+                                                                    </a>
+                                                                </div>
                                                             </div>
-                                                            <div class="user-info">
-                                                                <span class="tb-lead"><?php echo $row['user_fname']?> <?php echo $row['user_oname']?> <span class="dot dot-success d-md-none ml-1"></span></span>
-                                                                <span><?php echo $row['user_email']?></span>
-                                                            </div>
-                                                        </div>
-                                                    </div>
-                                                    <div class="nk-tb-col tb-col-mb">
-                                                        <span class="tb-amount"><?php echo $row['user_id']?></span></span>
-                                                    </div>
-                                                    <div class="nk-tb-col tb-col-md">
-                                                        <span>+254 <?php echo substr($row['phone_number'],1,3)?> <?php echo substr($row['phone_number'],4,3)?> <?php echo substr($row['phone_number'],7,3)?></span>
-                                                    </div>
-                                                    <div class="nk-tb-col tb-col-lg">
-                                                        <ul class="list-status">
-                                                            <li><em class="icon text-success ni ni-check-circle"></em> <span>Email</span></li>
-                                                            <li><em class="icon text-success ni ni-check-circle"></em> <span>KYC</span></li>
-                                                        </ul>
-                                                    </div>
-                                                    <div class="nk-tb-col tb-col-lg">
-                                                        <span><?php echo $row['added_on']?></span>
-                                                    </div>
-                                                    <div class="nk-tb-col tb-col-md">
-                                                        <span class="tb-status text-success"><?php echo $row['user_type']?></span>
-                                                    </div>
-                                                    <div class="nk-tb-col nk-tb-col-tools">
-                                                        <ul class="nk-tb-actions gx-1">
-                                                            <li class="nk-tb-action-hidden">
-                                                                <a href="/html/user-list-regular.html" class="btn btn-trigger btn-icon" data-toggle="tooltip" data-placement="top" title="Wallet">
-                                                                    <em class="icon ni ni-wallet-fill"></em>
-                                                                </a>
-                                                            </li>
-                                                            <li class="nk-tb-action-hidden">
-                                                                <a href="/html/user-list-regular.html" class="btn btn-trigger btn-icon" data-toggle="tooltip" data-placement="top" title="Send Email">
-                                                                    <em class="icon ni ni-mail-fill"></em>
-                                                                </a>
-                                                            </li>
-                                                            <li class="nk-tb-action-hidden">
-                                                                <a href="/html/user-list-regular.html" class="btn btn-trigger btn-icon" data-toggle="tooltip" data-placement="top" title="Suspend">
-                                                                    <em class="icon ni ni-user-cross-fill"></em>
-                                                                </a>
-                                                            </li>
-                                                            <li>
-                                                                <div class="drodown">
-                                                                    <a href="" class="dropdown-toggle btn btn-icon btn-trigger" data-toggle="dropdown"><em class="icon ni ni-more-h"></em></a>
-                                                                    <div class="dropdown-menu dropdown-menu-right">
-                                                                        <ul class="link-list-opt no-bdr">
-                                                                            <li><a href="/html/user-list-regular.html"><em class="icon ni ni-focus"></em><span>Quick View</span></a></li>
-                                                                            <li><a href="/html/user-list-regular.html"><em class="icon ni ni-eye"></em><span>View Details</span></a></li>
-                                                                            <li><a href="/html/user-list-regular.html"><em class="icon ni ni-repeat"></em><span>Transaction</span></a></li>
-                                                                            <li><a href="/html/user-list-regular.html"><em class="icon ni ni-activity-round"></em><span>Activities</span></a></li>
-                                                                            <li class="divider"></li>
-                                                                            <li><a href="/html/user-list-regular.html"><em class="icon ni ni-shield-star"></em><span>Reset Pass</span></a></li>
-                                                                            <li><a href="/html/user-list-regular.html"><em class="icon ni ni-shield-off"></em><span>Reset 2FA</span></a></li>
-                                                                            <li><a href="/html/user-list-regular.html"><em class="icon ni ni-na"></em><span>Suspend User</span></a></li>
-                                                                        </ul>
+                                                            <div class="dropdown-body dropdown-body-rg">
+                                                                <div class="row gx-6 gy-3">
+                                                                    <div class="col-6">
+                                                                        <div class="custom-control custom-control-sm custom-checkbox">
+                                                                            <input type="checkbox" class="custom-control-input" id="hasBalance">
+                                                                            <label class="custom-control-label" for="hasBalance"> Have Balance</label>
+                                                                        </div>
+                                                                    </div>
+                                                                    <div class="col-6">
+                                                                        <div class="custom-control custom-control-sm custom-checkbox">
+                                                                            <input type="checkbox" class="custom-control-input" id="hasKYC">
+                                                                            <label class="custom-control-label" for="hasKYC"> KYC Verified</label>
+                                                                        </div>
+                                                                    </div>
+                                                                    <div class="col-6">
+                                                                        <div class="form-group">
+                                                                            <label class="overline-title overline-title-alt">Role</label>
+                                                                            <select class="form-select">
+                                                                                <option value="any">Any Role</option>
+                                                                                <option value="investor">Investor</option>
+                                                                                <option value="seller">Seller</option>
+                                                                                <option value="buyer">Buyer</option>
+                                                                            </select>
+                                                                        </div>
+                                                                    </div>
+                                                                    <div class="col-6">
+                                                                        <div class="form-group">
+                                                                            <label class="overline-title overline-title-alt">Status</label>
+                                                                            <select class="form-select">
+                                                                                <option value="any">Any Status</option>
+                                                                                <option value="active">Active</option>
+                                                                                <option value="pending">Pending</option>
+                                                                                <option value="suspend">Suspend</option>
+                                                                                <option value="deleted">Deleted</option>
+                                                                            </select>
+                                                                        </div>
+                                                                    </div>
+                                                                    <div class="col-12">
+                                                                        <div class="form-group">
+                                                                            <button type="button" class="btn btn-secondary">Filter</button>
+                                                                        </div>
                                                                     </div>
                                                                 </div>
-                                                            </li>
-                                                        </ul>
-                                                    </div>
-                                                </div><!-- .nk-tb-item -->
-                                                <?php endforeach;?>
-                                            </div><!-- .nk-tb-list -->
-                                        </div><!-- .card-inner -->
-                                        
-                                    </div><!-- .card-inner-group -->
-                                </div><!-- .card -->
-                            </div><!-- .nk-block -->
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <!-- content @e -->
-
-            <div class="nk-add-product toggle-slide toggle-slide-right" data-content="addProduct" data-toggle-screen="any" data-toggle-overlay="true" data-toggle-body="true" data-simplebar>
-            <form id="addUser" name="addUser" method="post" action="/html/user-add-user.html">                     
-            <div class="nk-block-head">
-                                    <div class="nk-block-head-content">
-                                        <h5 class="nk-block-title">New User</h5>
-                                        <div class="nk-block-des">
-                                            <p>Add information and add new User.</p>
-                                        </div>
-                                    </div>
-                                </div><!-- .nk-block-head -->
-                                <div class="nk-block">
-                               
-                                    <div class="row g-3">
-                                        <div class="col-12">
-                                            <div class="form-group">
-                                                <label class="form-label" for="product-title">First Name</label>
-                                                <div class="form-control-wrap">
-                                                    <input type="text" class="form-control" id="txtFname" name="txtFname" placeholder="Franklin">
-                                                </div>
-                                            </div>
-                                        </div>
-
-                                        <div class="col-12">
-                                            <div class="form-group">
-                                                <label class="form-label" for="product-title">Other Name</label>
-                                                <div class="form-control-wrap">
-                                                    <input type="text" class="form-control" id="txtOname" name="txtOname" placeholder="Mutuma">
-                                                </div>
-                                            </div>
-                                        </div>
-
-                                        <div class="col-12">
-                                            <div class="form-group">
-                                                <label class="form-label" for="product-title">Email</label>
-                                                <div class="form-control-wrap">
-                                                    <input type="text" class="form-control" id="txtEmail" name="txtEmail" placeholder="info@camera20production.co.ke">
-                                                </div>
-                                            </div>
-                                        </div>
-                                        
-                                        <div class="col-md-6">
-                                            <div class="form-group">
-                                                <label class="form-label" for="sale-price">Id Number</label>
-                                                <div class="form-control-wrap">
-                                                    <input type="text" class="form-control" id="txtID" name="txtID" placeholder="12345678">
-                                                </div>
-                                            </div>
-                                        </div>
-                                        <div class="col-md-6">
-                                            <div class="form-group">
-                                                <label class="form-label" for="stock">User Type</label>
-                                                <div class="form-control-wrap">
-                                                            <select class="form-select" data-search="on" name="txtUT" id="txtUT" placeholder="Category">
-                                                                 
-                                                                <option value="User">User</option>
-                                                                <option value="Admin">Admin</option>
-                                                            </select>
+                                                            </div>
+                                                            <div class="dropdown-foot between">
+                                                                <a class="clickable" href="/html/user-list-regular.html">Reset Filter</a>
+                                                                <a href="/html/user-list-regular.html">Save Filter</a>
+                                                            </div>
+                                                        </div><!-- .filter-wg -->
+                                                    </div><!-- .dropdown -->
+                                                </li><!-- li -->
+                                                <li>
+                                                    <div class="dropdown">
+                                                        <a href="/html/user-list-regular.html" class="btn btn-trigger btn-icon dropdown-toggle" data-toggle="dropdown">
+                                                            <em class="icon ni ni-setting"></em>
+                                                        </a>
+                                                        <div class="dropdown-menu dropdown-menu-xs dropdown-menu-right">
+                                                            <ul class="link-check">
+                                                                <li><span>Show</span></li>
+                                                                <li class="active"><a href="/html/user-list-regular.html">10</a></li>
+                                                                <li><a href="/html/user-list-regular.html">20</a></li>
+                                                                <li><a href="/html/user-list-regular.html">50</a></li>
+                                                            </ul>
+                                                            <ul class="link-check">
+                                                                <li><span>Order</span></li>
+                                                                <li class="active"><a href="/html/user-list-regular.html">DESC</a></li>
+                                                                <li><a href="/html/user-list-regular.html">ASC</a></li>
+                                                            </ul>
                                                         </div>
-                                                    </div>
-                                            </div>
-                                        </div>
-
-                                        <div class="col-md-6">
-                                            <div class="form-group">
-                                                <label class="form-label" for="stock">Phone Number</label>
-                                                <div class="form-control-wrap">
-                                                    <input type="text" class="form-control" id="txtPN" name="txtPN" placeholder="0796096678">
-                                                </div>
-                                            </div>
-                                        </div>
-                                       
-                                        <div class="col-12">
-                                            <div class="upload-zone small bg-lighter my-2">
-                                                <div class="dz-message">
-                                                    <span class="dz-message-text">Profile Photo </span>
-                                                </div>
-                                            </div>
-                                        </div>
-                                        <div class="col-12">
-                                        <div class="form-group">
-                                            <button type="submit" class="btn btn-primary"><em class="icon ni ni-plus"></em><span>Add New</span></button>
-                                        </div></div>
-                                        </form>
-                                        
-                                    </div>
-                                </div><!-- .nk-block -->
-
+                                                    </div><!-- .dropdown -->
+                                                </li><!-- li -->
+                                            </ul><!-- .btn-toolbar -->
+                                        </div><!-- .toggle-content -->
+                                    </div><!-- .toggle-wrap -->
+                                </li><!-- li -->
+                            </ul><!-- .btn-toolbar -->
+                        </div><!-- .card-tools -->
+                    </div><!-- .card-title-group -->
+                    <div class="card-search search-wrap" data-search="search">
+                        <div class="card-body">
+                            <div class="search-content">
+                                <a href="/html/user-list-regular.html" class="search-back btn btn-icon toggle-search" data-target="search"><em class="icon ni ni-arrow-left"></em></a>
+                                <input type="text" class="form-control border-transparent form-focus-none" placeholder="Search by user or email">
+                                <button class="search-submit btn btn-icon"><em class="icon ni ni-search"></em></button>
                             </div>
+                        </div>
+                    </div><!-- .card-search -->
+                </div><!-- .card-inner -->
+                <div class="card-inner p-0">
+                    <div class="nk-tb-list nk-tb-ulist">
+                        <div class="nk-tb-item nk-tb-head">
+                            <div class="nk-tb-col nk-tb-col-check">
+                                <div class="custom-control custom-control-sm custom-checkbox notext">
+                                    <input type="checkbox" class="custom-control-input" id="uid">
+                                    <label class="custom-control-label" for="uid"></label>
+                                </div>
+                            </div>
+                            <div class="nk-tb-col"><span class="sub-text">User</span></div>
+                            <div class="nk-tb-col tb-col-mb"><span class="sub-text">ID Number</span></div>
+                            <div class="nk-tb-col tb-col-md"><span class="sub-text">Phone</span></div>
+                            <div class="nk-tb-col tb-col-lg"><span class="sub-text">Verified</span></div>
+                            <div class="nk-tb-col tb-col-lg"><span class="sub-text">Added On</span></div>
+                            <div class="nk-tb-col tb-col-md"><span class="sub-text">Account Type</span></div>
+                            <div class="nk-tb-col nk-tb-col-tools text-right">
+                                <div class="dropdown">
+                                    <a href="/html/user-list-regular.html" class="btn btn-xs btn-outline-light btn-icon dropdown-toggle" data-toggle="dropdown" data-offset="0,5"><em class="icon ni ni-plus"></em></a>
+                                    <div class="dropdown-menu dropdown-menu-xs dropdown-menu-right">
+                                        <ul class="link-tidy sm no-bdr">
+                                            <li>
+                                                <div class="custom-control custom-control-sm custom-checkbox">
+                                                    <input type="checkbox" class="custom-control-input" checked="" id="bl">
+                                                    <label class="custom-control-label" for="bl">ID Number</label>
+                                                </div>
+                                            </li>
+                                            <li>
+                                                <div class="custom-control custom-control-sm custom-checkbox">
+                                                    <input type="checkbox" class="custom-control-input" checked="" id="ph">
+                                                    <label class="custom-control-label" for="ph">Phone</label>
+                                                </div>
+                                            </li>
+                                            <li>
+                                                <div class="custom-control custom-control-sm custom-checkbox">
+                                                    <input type="checkbox" class="custom-control-input" id="vri">
+                                                    <label class="custom-control-label" for="vri">Verified</label>
+                                                </div>
+                                            </li>
+                                            <li>
+                                                <div class="custom-control custom-control-sm custom-checkbox">
+                                                    <input type="checkbox" class="custom-control-input" id="st">
+                                                    <label class="custom-control-label" for="st">Status</label>
+                                                </div>
+                                            </li>
+                                        </ul>
+                                    </div>
+                                </div>
+                            </div>
+                        </div><!-- .nk-tb-item -->
 
 
+                        <?php foreach ($user as $row) : ?>
+                            <div class="nk-tb-item">
+                                <div class="nk-tb-col nk-tb-col-check">
+                                    <div class="custom-control custom-control-sm custom-checkbox notext">
+                                        <input type="checkbox" class="custom-control-input" id="uid10">
+                                        <label class="custom-control-label" for="uid10"></label>
+                                    </div>
+                                </div>
+                                <div class="nk-tb-col">
+                                    <div class="user-card">
+                                        <div class="user-avatar bg-purple">
+                                            <span><?php echo substr($row['user_fname'], 0, 1) ?><?php echo substr($row['user_oname'], 0, 1) ?></span>
+                                        </div>
+                                        <div class="user-info">
+                                            <span class="tb-lead"><?php echo $row['user_fname'] ?> <?php echo $row['user_oname'] ?> <span class="dot dot-success d-md-none ml-1"></span></span>
+                                            <span><?php echo $row['user_email'] ?></span>
+                                        </div>
+                                    </div>
+                                </div>
+                                <div class="nk-tb-col tb-col-mb">
+                                    <span class="tb-amount"><?php echo $row['user_id'] ?></span></span>
+                                </div>
+                                <div class="nk-tb-col tb-col-md">
+                                    <span>+254 <?php echo substr($row['phone_number'], 1, 3) ?> <?php echo substr($row['phone_number'], 4, 3) ?> <?php echo substr($row['phone_number'], 7, 3) ?></span>
+                                </div>
+                                <div class="nk-tb-col tb-col-lg">
+                                    <ul class="list-status">
+                                        <li><em class="icon text-success ni ni-check-circle"></em> <span>Email</span></li>
+                                        <li><em class="icon text-success ni ni-check-circle"></em> <span>KYC</span></li>
+                                    </ul>
+                                </div>
+                                <div class="nk-tb-col tb-col-lg">
+                                    <span><?php echo $row['added_on'] ?></span>
+                                </div>
+                                <div class="nk-tb-col tb-col-md">
+                                    <span class="tb-status text-success"><?php echo $row['user_type'] ?></span>
+                                </div>
+                                <div class="nk-tb-col nk-tb-col-tools">
+                                    <ul class="nk-tb-actions gx-1">
+                                        <li class="nk-tb-action-hidden">
+                                            <a href="/html/user-list-regular.html" class="btn btn-trigger btn-icon" data-toggle="tooltip" data-placement="top" title="Wallet">
+                                                <em class="icon ni ni-wallet-fill"></em>
+                                            </a>
+                                        </li>
+                                        <li class="nk-tb-action-hidden">
+                                            <a href="/html/user-list-regular.html" class="btn btn-trigger btn-icon" data-toggle="tooltip" data-placement="top" title="Send Email">
+                                                <em class="icon ni ni-mail-fill"></em>
+                                            </a>
+                                        </li>
+                                        <li class="nk-tb-action-hidden">
+                                            <a href="/html/user-list-regular.html" class="btn btn-trigger btn-icon" data-toggle="tooltip" data-placement="top" title="Suspend">
+                                                <em class="icon ni ni-user-cross-fill"></em>
+                                            </a>
+                                        </li>
+                                        <li>
+                                            <div class="drodown">
+                                                <a href="" class="dropdown-toggle btn btn-icon btn-trigger" data-toggle="dropdown"><em class="icon ni ni-more-h"></em></a>
+                                                <div class="dropdown-menu dropdown-menu-right">
+                                                    <ul class="link-list-opt no-bdr">
+                                                        <li><a href="/html/user-list-regular.html"><em class="icon ni ni-focus"></em><span>Quick View</span></a></li>
+                                                        <li><a href="/html/user-list-regular.html"><em class="icon ni ni-eye"></em><span>View Details</span></a></li>
+                                                        <li><a href="/html/user-list-regular.html"><em class="icon ni ni-repeat"></em><span>Transaction</span></a></li>
+                                                        <li><a href="/html/user-list-regular.html"><em class="icon ni ni-activity-round"></em><span>Activities</span></a></li>
+                                                        <li class="divider"></li>
+                                                        <li><a href="/html/user-list-regular.html"><em class="icon ni ni-shield-star"></em><span>Reset Pass</span></a></li>
+                                                        <li><a href="/html/user-list-regular.html"><em class="icon ni ni-shield-off"></em><span>Reset 2FA</span></a></li>
+                                                        <li><a href="/html/user-list-regular.html"><em class="icon ni ni-na"></em><span>Suspend User</span></a></li>
+                                                    </ul>
+                                                </div>
+                                            </div>
+                                        </li>
+                                    </ul>
+                                </div>
+                            </div><!-- .nk-tb-item -->
+                        <?php endforeach; ?>
+                    </div><!-- .nk-tb-list -->
+                </div><!-- .card-inner -->
 
-                       
-<script>
-    $('#addUser').validate({
-        rules:{
-            txtFname: "required",
-            txtOname: "required",
-            txtEmail: "required",
-            txtID: "required",
-            txtUT: "required",
-            txtPN: "required",
-        },
-        messages:{
-            txtFname: "First name is required",
-            txtOname: "Other name is required",
-            txtEmail: "Email is required",
-            txtID: "User national ID isrequired",
-            txtUT: "User Type is required",
-            txtPN: "User phone number is required",
-        },
-        submitHandler: function(form){
-            var form_action = $("#addUser").attr("action");
-                    $.ajax({
-                        data: $('#addUser').serialize(),
-                        url: form_action,
-                        type: 'POST',
-                        dataType: 'json',
-                        beforeSend:function(){Swal.fire({icon: 'info',title: 'Loading...',showConfirmButton: false,timer: 3500})},
-                        success: function (res) {
-                            
-                            var $status =  + JSON.stringify(res.status);
-                            var $sts = 'false';
-                            if ($status < 1){
-                                Swal.fire({
-                                icon:'error',
-                                title: 'Ooops...',
-                                text: JSON.stringify(res.data.message)
-                                })
-                            }else{
-                                Swal.fire({
-                                icon:'success',
-                                title: 'Success',
-                                text: JSON.stringify(res.data.message)
-                                }).then(()=>{
-                                    window.location.href = "/html/user-list-regular.html";
-                                });
-                            }
-                            $('#loading').modal('hide');
+            </div><!-- .card-inner-group -->
+        </div><!-- .card -->
+    </div><!-- .nk-block -->
+</div><!-- .nk-content -->
 
-                        },
-                        error: function (data) {
-                            
-                            Swal.fire({
-                                icon:'error',
-                                title: 'Ooops...',
-                                text: "An error: "+JSON.stringify(data.responseText)+" has occured"
-                            })
-                        }
-            });
-        }
-    });
-</script>
+
+<!-- Load Modals -->
+<?= $this->include('user/user_modals') ?>
+<!-- End Load Modals -->
+
+
+<!-- content ends here -->
+<?= $this->endSection('cp') ?>
